bugs fixed and navbar improvement

+++ blogs/views.py
@@ -1,6 +1,9 @@
 # _*_ coding:utf-8 _*_
-from django.contrib.auth import authenticate,login
+from django.contrib import auth
+from django.contrib.auth import authenticate,login,logout
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+
 
 from django.views.decorators.csrf import csrf_exempt
 from django.shortcuts import render, redirect
@@ -19,10 +22,13 @@ class IndexView(generic.ListView):
     template_name = 'blogs/index.html'
     context_object_name = 'articles'
 
-    def get_queryset(self,**kwargs):
-        kwargs['isHome']=True
+    def get_queryset(self):
         return Article.objects.order_by('-edit_date')
 
+    def get_context_data(self, **kwargs):
+        kwargs['isHome'] = True
+        kwargs['isLogin'] = isLogin(self.request)
+        return super(IndexView, self).get_context_data(**kwargs)
 
 class CategoryView(generic.ListView):
     template_name = 'blogs/category.html'
@@ -32,10 +38,20 @@ class CategoryView(generic.ListView):
         kwargs['isCategory']=True
         return Article.objects.order_by('-pub_date')
 
+    def get_context_data(self, **kwargs):
+        kwargs['isCategory'] = True
+        kwargs['isLogin'] = isLogin(self.request)
+        return super(CategoryView, self).get_context_data(**kwargs)
+
 
 class ArticleView(generic.DetailView):
     model = Article
     template_name = 'blogs/article.html'
+
+    def get_context_data(self, **kwargs):
+        kwargs['isLogin'] = isLogin(self.request)
+        return super(ArticleView, self).get_context_data(**kwargs)
+
 
 
 class CreateView(generic.CreateView):
@@ -45,11 +61,17 @@ class CreateView(generic.CreateView):
 
     @csrf_exempt
     def form_valid(self, form):
+        if isLogin(self.request):
+            article = form.save(commit=False)
+            article.author = self.request.user
+            article.save()
+            return redirect("/blogs/")
+        else:
+            return redirect("/blogs/login")
 
-        article = form.save(commit=False)
-        article.author = self.request.user
-        article.save()
-        return redirect("/blogs/")
+    def get_context_data(self, **kwargs):
+        kwargs['isLogin'] = isLogin(self.request)
+        return super(CreateView, self).get_context_data(**kwargs)
 
 
 class SearchView(generic.ListView):
@@ -60,22 +82,33 @@ class SearchView(generic.ListView):
         keyword = self.request.GET.get('keyword')
         return Article.objects.filter(title__contains=keyword)
 
-
 class EditView(generic.DetailView):
     model = Article
     template_name = 'blogs/edit.html'
 
+    def get_context_data(self, **kwargs):
+        kwargs['isLogin'] = isLogin(self.request)
+        return super(EditView, self).get_context_data(**kwargs)
+
+@login_required
 def edit(request):
     article = Article.objects.get(pk=request.POST['pk'])
-    article.title = request.POST.get("title")
-    article.content = request.POST.get("content")
-    article.save()
-    return redirect("/blogs/article/"+request.POST['pk'])
+    if isAuthor(request,article):
+        article.title = request.POST.get("title")
+        article.content = request.POST.get("content")
+        article.save()
+        return redirect("/blogs/article/"+request.POST['pk'])
+    else:
+        return redirect("/blogs/login")
 
+@login_required
 def delete(request):
     article=Article.objects.get(pk=request.POST['pk'])
-    article.delete()
-    return redirect("/blogs/")
+    if isAuthor(request,article):
+        article.delete()
+        return redirect("/blogs")
+    else:
+        return redirect("/blogs/login")
 
 
 def login(request):
@@ -86,7 +119,7 @@ def login(request):
         password=request.POST['password']
         user=authenticate(username=username,password=password)
         if user is not None:
-            login(request,user)
+            auth.login(request,user)
             return redirect('/blogs')
         else:
             return redirect('/blogs/login')
@@ -104,6 +137,26 @@ def register(request):
             user=User.objects.create_user(username,email,password)
             login(request,user)
             return redirect('/blogs')
+
+
+def logout(request):
+    auth.logout(request)
+    return redirect('/blogs')
+
+
+def isLogin(request):
+    if request.user.is_authenticated():
+        return True
+    else:
+        return False
+
+def isAuthor(request,article):
+    if request.user.username==article.author:
+        return True
+    else:
+        return False
+
+
